Move base store to common
