+++ ui/src/user/UserStore.ts
@@ -1,4 +1,4 @@
-import {BaseStore} from '../stores/BaseStore';
+import {BaseStore} from '../common/BaseStore';
 import axios from 'axios';
 import * as config from '../config';
 import {action} from 'mobx';
